Refactor tokenizer pipeline
The first iteration of the tokenizer pipeline was pretty brittle and did
not allow for easy configuration.

A side effect was that raw tokens still got split on non-word
characters, which meant you couldn't do term filtering on strings like
'live_events', as that string would be tokenized as 'live', 'events'.

This PR refactors the pipeline to make it easy to pick and choose which
operations to apply in the pipeline, allowing for the term indexing to
skip tokenization while keeping the whitespace splitter and general
string cleanup.